Redesign into packages
