+++ data/tasks.txt
@@ -1,3 +1,4 @@
 T | 1 | eat
 T | 1 | sleep
 D | 0 | study | 2025-10-10
+T | 0 | cry
