+++ src/main/java/avocado/command/ExitCommand.java
@@ -1,3 +1,10 @@
+package avocado.command;
+
+import avocado.exception.AvocadoException;
+import avocado.storage.Storage;
+import avocado.task.TaskList;
+import avocado.ui.Ui;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws AvocadoException {
