+++ src/main/java/avocado/exception/AvocadoException.java
@@ -1,3 +1,5 @@
+package avocado.exception;
+
 public class AvocadoException extends Exception {
     public AvocadoException(String message) {
         super(message);
