+++ src/main/java/avocado/task/Event.java
@@ -1,8 +1,8 @@
-
+package avocado.task;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-class Event extends Task {
+public class Event extends Task {
     protected LocalDateTime from;
     protected LocalDateTime to;
