+++ src/main/java/avocado/task/TaskList.java
@@ -1,3 +1,4 @@
+package avocado.task;
 
 import java.util.ArrayList;
