+++ src/main/java/avocado/task/Todo.java
@@ -1,4 +1,6 @@
-class Todo extends Task {
+package avocado.task;
+
+public class Todo extends Task {
     public Todo(String description) {
         super(description);
     }
